Separate CONFIGURE_SUBMITTY, add defaults to data folder, update CI to work
