+++ .setup/ansible/roles/submitty_install/defaults/main.yml
@@ -29,3 +29,4 @@ submitty_install_ssl_key_file: /etc/ssl/private/submitty.key
 submitty_install_submitty_version: main
 submitty_install_github_url: https://github.com/Submitty/Submitty.git
 submitty_install_ssl_enabled: false
+submitty_install_config_location: /usr/local/submitty/GIT_CHECKOUT/Submitty/.setup/data/configs/
