+++ .setup/data/configs/autograding_workers.json
@@ -1,0 +1,17 @@
+{
+    "primary": {
+        "capabilities": [
+            "default",
+            "cpp",
+            "python",
+            "et-cetera",
+            "notebook",
+            "unsupported"
+        ],
+        "address": "localhost",
+        "username": "",
+        "num_autograding_workers": 5,
+        "enabled": true
+    }
+}
+
