+++ .setup/data/configs/submitty.json
@@ -1,0 +1,42 @@
+{
+  "submitty_install_dir": "/usr/local/submitty",
+  "submitty_repository": "/usr/local/submitty/GIT_CHECKOUT/Submitty",
+  "submitty_data_dir": "/var/local/submitty",
+  "autograding_log_path": "/var/local/submitty/logs/autograding",
+  "sys_admin_email": "sysadmin@example.com",
+  "sys_admin_url": "https://example.com",
+  "site_log_path": "/var/local/submitty/logs",
+  "submission_url": "http://localhost:1511",
+  "vcs_url": "",
+  "cgi_url": "http://localhost:1511/cgi-bin",
+  "websocket_port": 8443,
+  "institution_name": "",
+  "institution_homepage": "",
+  "timezone": "America/New_York",
+  "default_locale": "en_US",
+  "duck_special_effects": false,
+  "course_material_file_upload_limit_mb": "100",
+  "user_create_account": false,
+  "user_id_requirements": {
+    "any_user_id": true,
+    "require_name": false,
+    "min_length": 6,
+    "max_length": 25,
+    "name_requirements": {
+      "given_first": false,
+      "given_name": 2,
+      "family_name": 4
+    },
+    "require_email": false,
+    "email_requirements": {
+      "whole_email": false,
+      "whole_prefix": false,
+      "prefix_count": 6
+    },
+    "accepted_emails": [
+      "gmail.com"
+    ]
+  },
+  "worker": false
+}
+
