+++ .setup/data/configs/submitty_users.json
@@ -1,0 +1,3 @@
+{
+    "supervisor_user": "submitty"
+}
