+++ .setup/install_system.sh
@@ -75,7 +75,7 @@ echo "Running setup:"
 while :; do
     case $1 in
         --utm)
-            export UTM=1
+            # export UTM=1
             export DEV_VM=1
             echo "utm"
             ;;
@@ -107,6 +107,20 @@ fi
 if [ ${UTM} == 1 ]; then
     mkdir ${SUBMITTY_REPOSITORY}/.utm
 fi
+
+COMMAND=""
+if [ ${WORKER} == 1 ]; then
+    COMMAND=--worker
+fi
+
+if [ ${DEV_VM} == 1 ]; then
+    python3 ${CURRENT_DIR}/generate_configs.py ${COMMAND}
+fi
+
+if [ ${DEV_VM} == 0 ]; then
+    python3 ${COMMAND} ${CURRENT_DIR}/validate_configs.py
+fi
+
 
 if [ ${DEV_VM} == 1 ] && [ ${WORKER} == 0 ]; then
     # Setting it up to allow SSH as root by default
@@ -189,11 +203,7 @@ systemctl start submitty_daemon_jobs_handler
 systemctl start submitty_websocket_server
 systemctl start nullsmtpd
 cd ${SUBMITTY_INSTALL_DIR}" >> /root/.bashrc
-else
-    #TODO: We should get options for ./.setup/CONFIGURE_SUBMITTY.py script
-    :
 fi
-
 if [ ${WORKER} == 1 ]; then
     if [ ${VAGRANT} == 1 ]; then
         # Setting it up to allow SSH as root by default
@@ -348,6 +358,9 @@ echo "$gitconfig_content" > "$gitconfig_path"
 sudo chown "${DAEMON_USER}:${DAEMON_USER}" "$gitconfig_path"
 
 usermod -a -G docker "${DAEMON_USER}"
+
+python3 ${COMMAND} ${CURRENT_DIR}/set_config_permissions.py
+
 
 #################################################################
 # JAR SETUP
@@ -667,58 +680,10 @@ if [ ! -d "${clangsrc}" ]; then
     echo 'DONE PREPARING CLANG INSTALLATION'
 fi
 
-#################################################################
-# SUBMITTY SETUP
-#################
+# #################################################################
+# # SUBMITTY SETUP
+# #################
 echo Beginning Submitty Setup
-
-#If in worker mode, run configure with --worker option.
-if [ ${WORKER} == 1 ]; then
-    echo "Running configure submitty in worker mode"
-    if [ ${DEV_VM} == 1 ]; then
-        echo "submitty" | python3 ${SUBMITTY_REPOSITORY}/.setup/CONFIGURE_SUBMITTY.py --worker
-    else
-        python3 ${SUBMITTY_REPOSITORY}/.setup/CONFIGURE_SUBMITTY.py --worker
-    fi
-else
-    if [ ${DEV_VM} == 1 ]; then
-        # This should be set by setup_distro.sh for whatever distro we have, but
-        # in case it is not, default to our primary URL
-        if [ -z "${SUBMISSION_URL}" ]; then
-            SUBMISSION_URL='http://192.168.56.101'
-        fi
-        echo -e "/var/run/postgresql
-${DB_USER}
-${DATABASE_PASSWORD}
-${DB_COURSE_USER}
-${DB_COURSE_PASSWORD}
-America/New_York
-en_US
-100
-${SUBMISSION_URL}
-
-
-sysadmin@example.com
-https://example.com
-1
-submitty-admin
-y
-
-
-submitty@vagrant
-do-not-reply@vagrant
-localhost
-25
-" | python3 ${SUBMITTY_REPOSITORY}/.setup/CONFIGURE_SUBMITTY.py --debug --setup-for-sample-courses --websocket-port ${WEBSOCKET_PORT}
-
-        # Set these manually as they're not asked about during CONFIGURE_SUBMITTY.py
-        sed -i -e 's/"url": ""/"url": "ldap:\/\/localhost"/g' ${SUBMITTY_INSTALL_DIR}/config/authentication.json
-        sed -i -e 's/"uid": ""/"uid": "uid"/g' ${SUBMITTY_INSTALL_DIR}/config/authentication.json
-        sed -i -e 's/"bind_dn": ""/"bind_dn": "ou=users,dc=vagrant,dc=local"/g' ${SUBMITTY_INSTALL_DIR}/config/authentication.json
-    else
-        python3 ${SUBMITTY_REPOSITORY}/.setup/CONFIGURE_SUBMITTY.py
-    fi
-fi
 
 if [ ${WORKER} == 1 ]; then
    #Add the submitty user to /etc/sudoers if in worker mode.
@@ -734,8 +699,8 @@ fi
 
 # Create and setup database for non-workers
 if [ ${WORKER} == 0 ]; then
-    dbuser_password=`cat ${SUBMITTY_INSTALL_DIR}/.setup/submitty_conf.json | jq .database_password | tr -d '"'`
-    dbcourse_user_password=`cat ${SUBMITTY_INSTALL_DIR}/.setup/submitty_conf.json | jq .database_course_password | tr -d '"'`
+    dbuser_password=`cat ${SUBMITTY_INSTALL_DIR}/config/database.json | jq .database_password | tr -d '"'`
+    dbcourse_user_password=`cat ${SUBMITTY_INSTALL_DIR}/config/database.json | jq .database_course_password | tr -d '"'`
 
     # create the submitty_dbuser role in postgres (if it does not yet exist)
     # SUPERUSER privilege is required to use dblink extension (needed for data sync between master and course DBs).
