+++ .setup/testing/setup.sh
@@ -47,32 +47,16 @@ chown ${PHP_USER}:${PHP_GROUP} ${SUBMITTY_DATA_DIR}
 chmod -R 777 ${SUBMITTY_INSTALL_DIR}
 chmod -R 777 ${SUBMITTY_DATA_DIR}
 
-echo -e "localhost
-5432
-submitty_dbuser
-submitty_dbuser
-submitty_course_dbuser
-submitty_course_dbuser
-America/New_York
-en_US
-100
-http://localhost
+python3 ${SUBMITTY_REPOSITORY}/.setup/generate_configs.py --install-dir $SUBMITTY_INSTALL_DIR --data-dir $SUBMITTY_DATA_DIR
 
+ls -alF ${SUBMITTY_INSTALL_DIR}/config
+tmp=$(mktemp) && jq --arg v "localhost" '.database_host = $v' "$SUBMITTY_INSTALL_DIR/config/database.json" > "$tmp" && mv "$tmp" "$SUBMITTY_INSTALL_DIR/config/database.json"
+tmp=$(mktemp) && jq --arg v "DatabaseAuthentication" '.authentication_method = $v' "$SUBMITTY_INSTALL_DIR/config/database.json" > "$tmp" && mv "$tmp" "$SUBMITTY_INSTALL_DIR/config/database.json"
+tmp=$(mktemp) && jq --arg v "DatabaseAuthentication" '.authentication_method = $v' "$SUBMITTY_INSTALL_DIR/config/authentication.json" > "$tmp" && mv "$tmp" "$SUBMITTY_INSTALL_DIR/config/authentication.json"
+tmp=$(mktemp) && jq --arg v "http://localhost" '.submission_url = $v' "$SUBMITTY_INSTALL_DIR/config/submitty.json" > "$tmp" && mv "$tmp" "$SUBMITTY_INSTALL_DIR/config/submitty.json"
+tmp=$(mktemp) && jq --arg v "http://localhost/cgi-bin" '.cgi_url = $v' "$SUBMITTY_INSTALL_DIR/config/submitty.json" > "$tmp" && mv "$tmp" "$SUBMITTY_INSTALL_DIR/config/submitty.json"
 
-sysadmin@example.com
-https://example.com
-2
-y
-
-
-y
-
-
-submitty@vagrant
-do-not-reply@vagrant
-localhost
-25
-" | python3 ${SUBMITTY_REPOSITORY}/.setup/CONFIGURE_SUBMITTY.py --debug --setup-for-sample-courses --install-dir $SUBMITTY_INSTALL_DIR --data-dir $SUBMITTY_DATA_DIR
+python3 ${SUBMITTY_REPOSITORY}/.setup/set_config_permissions.py
 
 bash -c "echo 'export PATH=${PATH}' >> /home/${PHP_USER}/.profile"
 bash -c "echo 'export PATH=${PATH}' >> /home/${PHP_USER}/.bashrc"
